did refactoring around parsing codes and removed needless comments.
