+++ main/main.go
@@ -4,5 +4,5 @@ import "github.com/ryogrid/SamehadaDB/parser"
 
 func main() {
 	//fmt.Printf("hello world!")
-	parser.TestParsing()
+	parser.ProcessSQLStr()
 }
